CLC-5731 Allow OEC tasks to take department-filter and local-write arguments

+++ app/models/oec/task.rb
@@ -6,13 +6,14 @@ module Oec
       @log = []
       uid = Settings.oec.google.uid
       @remote_drive = GoogleApps::SheetsManager.new(uid, Settings.oec.google.marshal_dump)
-      @term_code = opts[:term_code]
+      @term_code = opts.delete :term_code
+      @opts = opts
       @tmp_path = Rails.root.join('tmp', 'oec')
     end
 
     def run
       log :info, "Starting #{self.class.name}"
-      run_internal
+      run_internal @opts
       true
     rescue => e
       log :error, "#{self.class.name} aborted with error: #{e.message}\n#{e.backtrace.join "\n\t"}"
@@ -24,19 +25,21 @@ module Oec
     private
 
     def create_folder(folder_name, parent=nil)
+      return if @opts[:local_write]
       if @remote_drive.find_folders_by_title(folder_name, folder_id(parent)).any?
-        raise RuntimeError, "Folder #{folder_name} with parent #{folder_title(parent)} already exists on remote drive"
+        raise RuntimeError, "Folder '#{folder_name}' with parent '#{folder_title(parent)}' already exists on remote drive"
       else
         create_folder_no_existence_check(folder_name, parent)
       end
     end
 
     def export_sheet(worksheet, dest_folder)
-      worksheet.export
-      log :debug, "Exported worksheet file #{worksheet.output_filename}"
-      upload_to_remote_drive(worksheet, worksheet.base_filename.chomp('.csv'), dest_folder)
-    ensure
-    #  File.delete csv.output_filename
+      if @opts[:local_write]
+        worksheet.export
+        log :debug, "Exported local worksheet file '#{worksheet.output_filename}'"
+      else
+        upload_to_remote_drive(worksheet, worksheet.base_filename.chomp('.csv'), dest_folder)
+      end
     end
 
     def find_or_create_folder(folder_name, parent=nil)
@@ -45,20 +48,20 @@ module Oec
 
     def create_folder_no_existence_check(folder_name, parent=nil)
       unless folder = @remote_drive.create_folder(folder_name, folder_id(parent))
-        raise RuntimeError, "Could not create folder #{folder_name} on remote drive"
+        raise RuntimeError, "Could not create folder '#{folder_name}' on remote drive"
       end
-      log :debug, "Created remote folder \"#{folder_name}\""
+      log :debug, "Created remote folder '#{folder_name}'"
       folder
     end
 
     def copy_file(file, dest_folder)
       if find_item(file.title, dest_folder)
-        raise RuntimeError, "File \"#{file.title}\" already exists in remote drive folder \"#{dest_folder.title}\"; could not copy"
+        raise RuntimeError, "File '#{file.title}' already exists in remote drive folder '#{dest_folder.title}'; could not copy"
       end
       if (@remote_drive.copy_item_to_folder(file, dest_folder.id))
-        log :debug, "Copied file \"#{file.title}\" to remote drive folder \"#{dest_folder.title}\""
+        log :debug, "Copied file '#{file.title}' to remote drive folder '#{dest_folder.title}'"
       else
-        raise RuntimeError, "Could not copy file \"#{file.title}\" to \"#{dest_folder.title}\""
+        raise RuntimeError, "Could not copy file '#{file.title}' to '#{dest_folder.title}'"
       end
     end
 
@@ -97,49 +100,49 @@ module Oec
 
     def upload_worksheet_headers(klass, dest_folder)
       worksheet = klass.new(@tmp_path)
-      begin
+      if @opts[:local_write]
         worksheet.export
-        log :debug, "Created header-only file #{worksheet.output_filename}"
+        log :debug, "Exported local header-only file #{worksheet.output_filename}"
+      else
         upload_to_remote_drive(worksheet, klass.name.demodulize.underscore, dest_folder)
-      ensure
-        File.delete worksheet.output_filename
       end
     end
 
     def upload_to_remote_drive(worksheet, title, folder)
       if @remote_drive.find_items_by_title(title, parent_id: folder.id).any?
-        raise RuntimeError, "File \"#{title}\" already exists in remote drive folder \"#{folder.title}\"; could not upload"
+        raise RuntimeError, "Item '#{title}' already exists in remote drive folder '#{folder.title}'; could not upload sheet"
       end
-      path = worksheet.output_filename
       if (!@remote_drive.upload_worksheet(title, '', worksheet, folder.id))
-        raise RuntimeError, "File #{path} could not be uploaded as sheet '#{title}' to remote drive folder \"#{folder.title}\""
+        raise RuntimeError, "Sheet '#{title}' could not be uploaded to remote drive folder '#{folder.title}'"
       end
-      log :debug, "Uploaded file #{path} as sheet '#{title}' to remote drive folder '#{folder.title}'"
+      log :debug, "Uploaded sheet '#{title}' to remote drive folder '#{folder.title}'"
     end
 
     def upload_file(path, remote_name, type, folder)
       if @remote_drive.find_items_by_title(remote_name, parent_id: folder.id).any?
-        raise RuntimeError, "File \"#{remote_name}\" already exists in remote drive folder \"#{folder.title}\"; could not upload"
+        raise RuntimeError, "Item '#{remote_name}' already exists in remote drive folder '#{folder.title}'; could not upload"
       end
       if (!@remote_drive.upload_file(remote_name, '', folder.id, type, path.to_s))
-        raise RuntimeError, "File #{path} could not be uploaded to remote drive folder \"#{folder.title}\""
+        raise RuntimeError, "File #{path} could not be uploaded to remote drive folder '#{folder.title}'"
       end
-      log :debug, "Uploaded file #{path} to remote drive folder \"#{folder.title}\""
+      log :debug, "Uploaded file #{path} to remote drive folder '#{folder.title}'"
     end
 
     def write_log
       if (term_folder = find_folder @term_code) && (reports_folder = find_folder('reports', term_folder))
         reports_today = find_or_create_folder(datestamp, reports_folder)
         log_name = "#{timestamp}_#{self.class.name.demodulize.underscore}.log"
-        begin
-          File.open(@tmp_path.join(log_name), 'wb') { |f| f.puts @log }
-          upload_file(@tmp_path.join(log_name), log_name, 'text/plain', reports_today)
-        ensure
-          File.delete @tmp_path.join(log_name)
+        File.open(@tmp_path.join(log_name), 'wb') { |f| f.puts @log }
+        unless @opts[:local_write]
+          begin
+            upload_file(@tmp_path.join(log_name), log_name, 'text/plain', reports_today)
+          ensure
+            File.delete @tmp_path.join(log_name)
+          end
         end
       end
     rescue => e
-      logger.error "Could not upload log: #{e.message}\n#{e.backtrace.join "\n\t"}"
+      logger.error "Could not write log: #{e.message}\n#{e.backtrace.join "\n\t"}"
     end
   end
 end
